requirements.txt
asgiref==3.8.1
boto3==1.36.16
botocore==1.36.16
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.4
django-social-links-field==0.1.4
django-storages==1.14.4
djangorestframework==3.15.2
firebase-admin==6.6.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
msgpack==1.1.0
packaging==24.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
rsa==4.9
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
whitenoise==6.8.2
